Files modified: Calendar.js, UserService.cs
Fix: Set the full calendar all day option if applicable. Modified the end time passed to a full day event to be a day later so the full calendar ui picks it up. As this was the designed functionality of full calendar and all day events to cut off the previous day. Modified the change event date as there was an issue with drag and drop of all day events and the end time changing.

git-svn-id: https://teamsupport.svn.beanstalkapp.com/main/trunk@3394 02885c34-e171-44ab-8435-42fb7a3c3aaf