[Ticket: 36620] Api Limit Count Exclude limit exceeded error
The API will exclude the exceeded limit errors (403) from the api call count itself.
Also the labels in the Admin -> Integration -> TeamSupport API section were updated to make sure that it is understood that the limitation is per 24hrs (from the time an api call is received back 24hrs), NOT daily (midnight to midnight).