If the user is *not* logged into TS then the images and wikis won't show (unless the org has the allow unauthorized ON). I found some issues when opening Tickets (with images) in some email clients (gmail in mobile vs browser), and outlook. I explained (or tried) this to Diego. I found this website that kind of explains this: https://litmus.com/blog/gmail-adds-image-caching-what-you-need-to-know
Also, per Diego, "Let's start by locking everything down, and when things become unusable, we bring those up
And refine the plan".

I'm afraid that this will be the case and we will get a lot of complains or confused users on this. Specially because you HAVE to be logged into TS if you want to see any images or wikis.